Pull method up, eliminate refs to ServiceApp.index_controller
